+++ src/parser/model.ts
@@ -3,10 +3,12 @@ import * as vscode from 'vscode';
 /** Symbols and structures parsed from a file */
 export interface Script {
     funcDefs: FuncDef[];
-    funcRefs: FuncCall[];
+    funcRefs: FuncRef[];
     labels: Label[];
     variables: Variable[];
     blocks: Block[];
+    /** Paths to scripts included within the current one */
+    includedPaths: string[];
 }
 
 export interface Variable {
@@ -27,7 +29,7 @@ export class Label {
     ) {}
 }
 
-export class FuncCall {
+export class FuncRef {
     constructor(
         public name: string,
         public document: vscode.TextDocument,