+++ src/providers/ahkRenameProvider.ts
@@ -1,5 +1,7 @@
 import * as vscode from 'vscode';
 import { Parser } from '../parser/parser';
+import { getFuncDefByName } from '../parser/parser.utils';
+import { ConfigKey, Global } from '../common/global';
 
 export class AhkRenameProvider implements vscode.RenameProvider {
     async provideRenameEdits(
@@ -48,7 +50,8 @@ export class AhkRenameProvider implements vscode.RenameProvider {
         const wordRange = document.getWordRangeAtPosition(position);
         const word = document.getText(wordRange);
 
-        const funcDef = await Parser.getFuncDefByName(document, word);
+        const newSearch = Global.getConfig<boolean>(ConfigKey.funcDefSearch);
+        const funcDef = getFuncDefByName(document.uri.path, word, newSearch);
         if (funcDef) {
             return wordRange;
         }