+++ src/providers/defProvider.ts
@@ -1,8 +1,10 @@
 import * as vscode from 'vscode';
 import { Parser } from '../parser/parser';
-import { resolveIncludedPath } from './defProvider.utils';
+import { resolveIncludedPath } from '../common/utils';
 import { Out } from '../common/out';
 import { stat } from 'fs/promises';
+import { getFuncDefByName } from '../parser/parser.utils';
+import { ConfigKey, Global } from '../common/global';
 
 export class DefProvider implements vscode.DefinitionProvider {
     public async provideDefinition(
@@ -35,7 +37,10 @@ export class DefProvider implements vscode.DefinitionProvider {
             )
         ) {
             Out.debug(`${funcName} calling getFuncDefByName for word: ${word}`);
-            const funcDef = await Parser.getFuncDefByName(document, word);
+            const newSearch = Global.getConfig<boolean>(
+                ConfigKey.funcDefSearch,
+            );
+            const funcDef = getFuncDefByName(docPath, word, newSearch);
             Out.debug(`${funcName} funcDef.name: ${funcDef?.name}`);
             if (funcDef) {
                 return new vscode.Location(