+++ changelog.md
@@ -5,6 +5,11 @@
 -   Improve extension performance when loading AHK v1 scripts ([PR #615](https://github.com/mark-wiemer/ahkpp/pull/615))
 -   Change `Method` to `Function` in document symbols and internal references. This changes the document outline slightly, but should not change most theme icons and is more correct. ([PR #620](https://github.com/mark-wiemer/ahkpp/pull/620)).
 
+### v1-exclusive changes
+
+-   Have scripts referenced by `#include` bypass the `exclude` setting ([PR #623](https://github.com/mark-wiemer/ahkpp/pull/623))
+-   Add experimental `funcDefSearch` setting that improves function definition resolution ([PR #623](https://github.com/mark-wiemer/ahkpp/pull/623))
+
 ## 6.5.0 - 2025-02-23 💞
 
 ### Debugger